camera_extension: callback deadlock prevention
BZ: 59533

In libcamera2, all callbacks can't be flushed completely thread
safely in every situation.

Therefore we take the same approach as exists for the Android
takePicture API, which is that the application may not start preview
before the jpeg callback has returned. In our case, we mandate that
all jpeg callbacks have returned in the cases where multiple
takePicture calls are issued (burst, bracketing, HDR). This will
ensure picture callbacks are not left for flushing.

Change-Id: I5d860369252cf7d50a1cecf333304e8491fc562c
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/68371
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>