[PORT FROM MAIN] camera_extension: added ULL snapshot ID to the ULL data callback
BZ: 96404

The Ultra-low light functionality requires the shot to be identified
for the application-side processing. When the ULL shot is to be taken,
the next ULL snapshot identifier is first provided via a notify callback
to the application. Later the same ID will be provided with the snapshot
data. This patch adds the ID to the snapshot data callback.

Depends on changes:
libcamera2: I1b287874f85c3172acea53b51a2629f11dccc856

Change-Id: I900df84f101e6d06308abf01bf2eb3e495da5fc6
Orig-Change-Id: I9e2c53223f429e5b436e43968b78be05e253f799
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>