camera_extension: added ULL snapshot ID to the ULL data callback
BZ: 96404

The Ultra-low light functionality requires the shot to be identified
for the application-side processing. When the ULL shot is to be taken,
the next ULL snapshot identifier is first provided via a notify callback
to the application. Later the same ID will be provided with the snapshot
data. This patch adds the ID to the snapshot data callback.

Depends on changes:
libcamera2: I1b287874f85c3172acea53b51a2629f11dccc856

Change-Id: I9e2c53223f429e5b436e43968b78be05e253f799
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/98852
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>