camera_extension: Add low priority support with Intel Cam Extension
The SmartTimeOut application is using the front camera and face
detection to determine whether a user is actually using the phone or
not. The viewfinder is not visible during this use case.
However, when the user then tries to open the normal camera
application while this is happening, the opening will fail.

Add support in the CameraService Layer for a low priority camera
instance that can be kicked out immediately when the user starts
any camera activity.

Depends on change-Id: I96931d44451f28651ea3cf31e94e4dd996269d07

Change-Id: I4221065e601d0bcd05d9c12ea8508355ef3ed59d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-8557
Signed-off-by: Soren Friis <soren.friis@intel.com>
Reviewed-on: https://android.intel.com:443/305446