camera_extension: add dual camera mode settings Java API
BZ: 192393

Issue:
for the Kevlar project which is based on the MOFD, it has totally 3 sensors.
the back sensor is ov8858 (raw)
the front sensors are imx132 which is raw sensor
and the other one is ov680 which is soc.
and the ov680 is one special sensor which is used to measure the depth.
but the Google architecture just support 2 sensors, 0 is back, 1 is front.
we need to add dual camera mode settings Java API
to use the depth camera (ov680) or normal camera (imx132).

Solution:
before the front camera is opened,
the "depth" or "normal" information must be set to camera HAL firstly.
so it means that the dual camera mode must be set in the 0 camera.
then we could know which front camera to open,
the normal front camera or the extended camera.

this patch is used to add the dual camera mode settings Java API

Change-Id: I5a7a8af71324796df3edd716bbfbdf61bbfd2309
Signed-off-by: Yang Liang <liang.l.yang@intel.com>