Fix DepthFrameReader to handle the case of RGBA8888 (without changing framework code)
In case of RGBA8888 the framework will allocate the buffer to be of format IMPLEMENTATION_DEFINED
The DepthFrameReader is changed so that in such case it will receive the original reader format and
ignore the mismatch between buffer and reader original format.
Also in functions that query the buffer format, they are fixed to consider this special case also and
to use the original format instead of the fake one.

Change-Id: I80d273f606ca92879ade491ad606f5c94d084ea1
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26422
Reviewed-on: https://android.intel.com/310382
Reviewed-by: Greenberg, Aviv D <aviv.d.greenberg@intel.com>
Tested-by: Greenberg, Aviv D <aviv.d.greenberg@intel.com>
Reviewed-on: https://android.intel.com:443/345554