add DepthCameraCpatureSessionConfiguration Class
To allow usage of standard formats like RGB for depth preview, the API informs
the HAL using binder of special configuration of the stream.
This solution is not final, as it assumes that the order of the streams sent
to the hal is the same as the list of surfaces sent to CreateCaptureSession.

Change-Id: Ibbbd0c56594d92173764f006c14e61491e528b4d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26422
Reviewed-on: https://android.intel.com/326842
Reviewed-by: Greenberg, Aviv D <aviv.d.greenberg@intel.com>
Tested-by: Greenberg, Aviv D <aviv.d.greenberg@intel.com>
Reviewed-on: https://android.intel.com:443/345565