Support multi stage stream configuration
Instead of using the framework to configure special usage bits
We use hal camera service binder, to be able to control the order of the streams
received by the hal the procedure now involves several configure_output calls, based
on how many types of "special streams" (need to set usage bits) we have.
Finally a call of the createCaptureSession will be called with the user's handler and callback.

Change-Id: I8e39a07d7e05f9fb9211c9ad6387a498932805d2
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26422
Reviewed-on: https://android.intel.com/337173
Reviewed-by: Greenberg, Aviv D <aviv.d.greenberg@intel.com>
Tested-by: Greenberg, Aviv D <aviv.d.greenberg@intel.com>
Reviewed-on: https://android.intel.com:443/345582