camera_extension: add Acc client and service
1. add the Acc client and service.
2. the Acc client and service are in two processes
the client is in the calling process,
the service is in the media server process.
3. right now, the service just provides acceleration from IPU.
4. the libiacp.so is moved to the service,
the original design is that the libiacp.so is in the JNI layer.
in the libiacp.so, it will malloc memory in it.
the memory will be passed to the driver.
but the malloc memory couldn't be shared between processes,
so the libiacp.so is moved to the service.
5. in the service, it registers function pointers to libiacp.
the libiacp will call the callback functions,
6. in the callback functions,
the libacc functions will be called.
the libacc will call Acc V4l2 IOCTRL.

TODO: add the interfaces for general computing by using GPU.

Change-Id: Ia4234c9ac216881007d610b7af00a9135d6fbeb8
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-30614
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://icggerrit.ir.intel.com/47187
Reviewed-by: Jingzhi Yu <jingzhi.yu@intel.com>
Reviewed-by: Gang A Hu <gang.a.hu@intel.com>
Validation-Android: Wei Tao <wei.tao@intel.com>
Reviewed-by: Bin B Yang <bin.b.yang@intel.com>
Build: IMIN-legacy (SC)
Reviewed-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Approver: Jingzhi Yu <jingzhi.yu@intel.com>
Integrator: Wei Tao <wei.tao@intel.com>
Build: Wei Tao <wei.tao@intel.com>
Reviewed-by: Wei Tao <wei.tao@intel.com>
Maintainer: Wei Tao <wei.tao@intel.com>